Add hook to pass certain extra field data from cw_requests to cw_wikis (#579)
In the future, this will make us be able to remove the `addNewRow()`
method from RemoteWikiFactory and be able to migrate things we use in
mw-config or MirahezeMagic like `primary-domain` and `mediawiki-version`
from our own custom table fields we added to the `cw_wikis` table to be
natively possible in CreateWiki using the hook and the
natively-available `wiki_extra` field.

It will also allow us to remove the fake variable in WikiDiscover,
`$wgWikiDiscoverDescription` to actually make it not have to just read
fake variable from ManageWikSettings and instead be possible by just
pulling the description from this `wiki_extra` field.

The other benefit this will bring will also be allowing to select the
primary domain (which we use a custom field and hook in ManageWiki for)
directly from RequestWiki without having to rely on our custom fields or
an additional custom field in `cw_requests`, so this will allow us to
bring selecting the primary domain directly from RequestWiki as well.

This has many more benefits for external use as well, making CreateWiki
much more expandable utilizing hooks rather than custom very hacky
overrides.