Avoid regenerating dblists more than once per server (#788)
Before this change, if a global dblist update was triggered, each
request to an individual server would trigger a new dblist update until
the first dblist update on that server was done. This was because the
server would check the local dblist on each request to figure out
whether it is outdated, but if a slow dblist update was already in
progress, it would start an additional one for each request, since the
first one hadn't written the new timestamp to the file yet.

Now, when a server updates its dblists due to a global update, we store
the current timestamp in the local server cache. This way, the currently
ongoing update essentially places a lock which prevents any other dblist
updates on the same server until a new global update is triggered.